+++ rules-tests/Php82/Rector/FuncCall/Utf8DecodeEncodeToMbConvertEncodingRector/Fixture/fixture.php.inc
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Rector\Tests\Php82\Rector\FuncCall\Utf8DecodeToMbConvertEncodingRector\Fixture;
+namespace Rector\Tests\Php82\Rector\FuncCall\Utf8DecodeEncodeToMbConvertEncodingRector\Fixture;
 
 $utf8_string = "\x5A\x6F\xC3\xAB";
 utf8_decode($utf8_string);
@@ -13,7 +13,7 @@ utf8_decode($utf8_string);
 
 declare(strict_types=1);
 
-namespace Rector\Tests\Php82\Rector\FuncCall\Utf8DecodeToMbConvertEncodingRector\Fixture;
+namespace Rector\Tests\Php82\Rector\FuncCall\Utf8DecodeEncodeToMbConvertEncodingRector\Fixture;
 
 $utf8_string = "\x5A\x6F\xC3\xAB";
 mb_convert_encoding($utf8_string, 'ISO-8859-1');
