+++ rules-tests/Php82/Rector/FuncCall/Utf8DecodeEncodeToMbConvertEncodingRector/Fixture/skip_different_function_name.php.inc
@@ -1,0 +1,7 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\Tests\Php82\Rector\FuncCall\Utf8DecodeEncodeToMbConvertEncodingRector\Fixture;
+
+strlen('test');
