--- rules-tests/Php82/Rector/FuncCall/Utf8DecodeToMbConvertEncodingRector/Fixture/skip_different_function_name.php.inc
@@ -1,7 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Rector\Tests\Php82\Rector\FuncCall\Utf8DecodeToMbConvertEncodingRector\Fixture;
-
-strlen('test');
