--- rules-tests/Php82/Rector/FuncCall/Utf8DecodeToMbConvertEncodingRector/config/configured_rule.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-use Rector\Config\RectorConfig;
-use Rector\Php82\Rector\FuncCall\Utf8DecodeToMbConvertEncodingRector;
-
-return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->rule(Utf8DecodeToMbConvertEncodingRector::class);
-};
